ci: cleanup and update ci
We factor out shared parts of the CI script into separate actions, add
visual debt to the main CI script so its better when we hide them
elsewhere.

The documentation and validation jobs use the same scripts, they were
essentially copy-pasted many moons ago.

I've updated many of the comments to make sense again.

The Coq versions are controlled from a single place at the top. There is
a dedicated local GitHub action that does the smart thing and selects
the correct version in various jobs. This will make maintaining the
versions simpler and more straight forward.

I've added workarounds so that we build correctly with Rocq 9.0 since
the compatability shims are sometimes not available.

Signed-off-by: Ali Caglayan <alizter@gmail.com>